requirements.txt
streamlit==1.15.0
openai==0.27.0
azure-ai-openai==1.0.0
azure-core==1.25.0
python-dotenv==0.19.2
